Move keys to traits
