+++ sbt-converter/src/main/scala/org/scalablytyped/converter/plugin/ScalablyTypedConverterGenSourcePlugin.scala
@@ -16,30 +16,9 @@ import scala.util.Try
 object ScalablyTypedConverterGenSourcePlugin extends AutoPlugin {
   override def requires = ScalablyTypedPluginBase && ScalaJSBundlerPlugin
 
-  object autoImport {
-    sealed trait SourceGenMode
-    object SourceGenMode {
-      case object ResourceGenerator extends SourceGenMode
-      case class Manual(toDir: File, overrideToDir: Map[String, File] = Map.empty) extends SourceGenMode
-    }
-
-    val stSourceGenMode = settingKey[SourceGenMode]("Whether to run automatically as source generator or manually")
-
-    /**
-      * A list of library names you don't care too much about.
-      * The idea is that we can limit compile time (by a lot!)
-      */
-    val stMinimize = settingKey[Selection[String]]("Specify which libraries you want minimized")
-
-    /**
-      * If you care about a small set of specific things from a library you can explicitly say you want that.
-      * Examples:
-      * - `angularCommon.mod.AsyncPipe`
-      * - `std.console`
-      */
-    val stMinimizeKeep = settingKey[List[String]]("a list of things you want to keep from minimized libraries")
-
-    val stImport = taskKey[Seq[File]]("Imports all the bundled npm and generates bindings")
+  object autoImport extends GenSourceKeys {
+    type SourceGenMode = plugin.SourceGenMode
+    val SourceGenMode: plugin.SourceGenMode.type = plugin.SourceGenMode
   }
 
   override lazy val projectSettings: scala.Seq[Def.Setting[_]] = {
